Seperate DgraphBundledHelper to its own file
